+++ src/inspect_evals_lint/checks/__init__.py
@@ -9,6 +9,7 @@ from inspect_evals_lint.checks.best_practices import (
 from inspect_evals_lint.checks.code_quality import (
     check_private_api_imports,
     check_score_constants,
+    check_unscored_reason,
 )
 from inspect_evals_lint.checks.dependencies import check_external_dependencies
 from inspect_evals_lint.checks.file_structure import (
@@ -54,6 +55,7 @@ __all__ = [
     "check_task_overridable_defaults",
     "check_tests_exist",
     "check_tests_init",
+    "check_unscored_reason",
     "get_eval_path",
     "get_test_path",
 ]